README
Dining Philosopher’s problem:
This problem states that there are K number of philosophers sitting around a circular table with one chopstick placed between each pair of philosophers . The philosopher will be able to eat if he can pickup two chopsticks that are adjacent to the philosopher. This problem deals with the allocation of limited resources. 
Write Java Multithreading program to solve the Dining Philosopher’s problem.